split db manager into db manager and file manager
